move tracking node location
