--- src/tracking_video_node/package.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>tracking_video_node</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="ethan.present2005@gmail.com">ethanpresent</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>rclcpp</depend>
-  <depend>vision_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>cv_bridge</depend>
-  <depend>opencv2</depend>
-  <depend>message_filters</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
